LayoutTests/platform/wpe/fast/speechsynthesis/speech-synthesis-gc-utterance-crash-expected.txt

CONSOLE MESSAGE: TypeError: window.internals.enableMockSpeechSynthesizer is not a function. (In 'window.internals.enableMockSpeechSynthesizer()', 'window.internals.enableMockSpeechSynthesizer' is undefined)
FAIL successfullyParsed should be true. Was false.
Some tests failed.

TEST COMPLETE
FAIL successfullyParsed should be true. Was false.
Some tests failed.

TEST COMPLETE
